+++ packages/environment/src/deployments/ethereum.ts
@@ -124,6 +124,7 @@ export default defineDeployment<Deployment.ETHEREUM>({
         BalancerV2LiquidityAdapter: "0xab5da4aa08b56c7e5a9d5d8a5ff19cf09a88c305",
         BalancerV2StablePoolPriceFeed: "0x438254d66e2bc576819a76a0ecb11fe41195d39f",
         BalancerV2WeightedPoolPriceFeed: "0x083b8f465bba2bb274e48387e3f9c56323341286",
+        ChainlinkLikeWstethPriceFeed: "0x92829c41115311ca43d5c9f722f0e9e7b9fcd30a",
         CompoundAdapter: "0x6ce8095a692aff6698c3aa8593be3976b6b8743d",
         CompoundDebtPositionLib: "0xc8e0c169dae4cdfd2f51e7494d2b5459f9314ee9",
         CompoundDebtPositionParser: "0x8bfbd013020c4d26d043183eff600ea3b36b4afb",